refactor: AI 코드 리뷰 피드백 2차 반영
- useMarkdownImage 훅에 onError 콜백 추가
- CLAUDE.md에 인라인 스타일 예외 문서화 (@dnd-kit, 동적 width/height)
- ProjectEditor/ExperienceEditor 확장 상태를 인덱스 기반에서 ID 기반으로 변경
- ProjectEditor의 SortableItem 컴포넌트들을 별도 파일로 분리
